fix "cross entropy" typo
