+++ legacy/docs/configs/solver_group.md
@@ -153,8 +153,8 @@ L2正则化系数
 
 ## `loss`
 
-训练时选择的损失函数， 支持`softmax_loss(sotfmax with cross entroy loss)`,
-`dice_loss(dice coefficient loss)`, `bce_loss(binary cross entroy loss)`三种损失函数。
+训练时选择的损失函数， 支持`softmax_loss(sotfmax with cross entropy loss)`,
+`dice_loss(dice coefficient loss)`, `bce_loss(binary cross entropy loss)`三种损失函数。
 其中`dice_loss`和`bce_loss`仅在两类分割问题中适用，`softmax_loss`不能与`dice_loss`
 或`bce_loss`组合，`dice_loss`可以和`bce_loss`组合使用。使用示例如下：
